Add moduleNameMapper and other feedback
