+++ __mocks__/@react-native-async-storage/async-storage.js
@@ -1,2 +1,1 @@
-import AsyncStorageMock from "@react-native-async-storage/async-storage/jest/async-storage-mock"
-export default AsyncStorageMock
+export { AsyncStorageMock as default } from "@react-native-async-storage/async-storage/jest/async-storage-mock"
