+++ __mocks__/react-navigation-native.js
@@ -1,0 +1,10 @@
+jest.mock("@react-navigation/native", () => {
+  const actualNav = jest.requireActual("@react-navigation/native")
+  return {
+    ...actualNav,
+    useNavigation: () => ({
+      addListener: jest.fn(),
+      navigate: jest.fn(),
+    }),
+  }
+})
