+++ jest.config.js
@@ -12,6 +12,12 @@ module.exports = {
   },
   testRegex: "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
   moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+  rootDir: ".",
+  moduleNameMapper: {
+    "^@app/(.*)$": ["<rootDir>app/$1"],
+    "^@mocks/(.*)$": ["<rootDir>__mocks__/$1"],
+    "^@node_modules/(.*)$": ["<rootDir>node_modules/$1"],
+  },
   transformIgnorePatterns: [
     "node_modules/(?!(react-native" +
       "|@react-native" +
